# backend/openapi.yaml
openapi: 3.1.0
info:
  title: NoesisForge API
  description: |
    NoesisForge backend API documentation. This API provides authentication, user management, and role-based access control functionality.
    
    ## Authentication
    This API uses JWT (JSON Web Token) for authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per IP address.
    
  version: 1.0.0
  contact:
    name: NoesisForge API Support
    email: support@noesisforge.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.noesisforge.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        username:
          type: string
          description: Unique username
          minLength: 3
          maxLength: 50
        name:
          type: string
          description: Full name
          minLength: 2
          maxLength: 100
        avatar:
          type: string
          description: Avatar URL
          nullable: true
        bio:
          type: string
          description: User biography
          nullable: true
        status:
          type: string
          enum: [active, pending, suspended]
          description: User account status
        emailVerified:
          type: boolean
          description: Email verification status
        emailVerifiedAt:
          type: string
          format: date-time
          nullable: true
          description: Email verification timestamp
        lastLogin:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        roleID:
          type: string
          format: uuid
          description: Assigned role ID
        role:
          $ref: '#/components/schemas/Role'
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique role identifier
        name:
          type: string
          description: Role name (system identifier)
        displayName:
          type: string
          description: Human-readable role name
        description:
          type: string
          description: Role description
        isDefault:
          type: boolean
          description: Is this the default role for new users
        isSystem:
          type: boolean
          description: Is this a system role (cannot be deleted)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique permission identifier
        name:
          type: string
          description: Permission name (system identifier)
        displayName:
          type: string
          description: Human-readable permission name
        description:
          type: string
          description: Permission description
        category:
          type: string
          description: Permission category
        isSystem:
          type: boolean
          description: Is this a system permission
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Refresh token for obtaining new access tokens
        tokenType:
          type: string
          example: Bearer
          description: Token type
        expiresIn:
          type: integer
          description: Access token expiration time in seconds
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (optional)
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - name
        - password
        - passwordConfirm
      properties:
        email:
          type: string
          format: email
          description: User email address
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Unique username
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Full name
        password:
          type: string
          minLength: 8
          description: Password
        passwordConfirm:
          type: string
          description: Password confirmation
    LoginRequest:
      type: object
      required:
        - password
      properties:
        email:
          type: string
          format: email
          description: User email (either email or username required)
        username:
          type: string
          description: Username (either email or username required)
        password:
          type: string
          description: Password
        remember:
          type: boolean
          description: Remember login
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Full name
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Username
        bio:
          type: string
          description: User biography
        avatar:
          type: string
          description: Avatar URL
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          description: Current password
        newPassword:
          type: string
          minLength: 8
          description: New password
...
paths:
...
  /auth/register:
    post:
      summary: Register new user
      description: Creates a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              username: johndoe
              name: John Doe
              password: securepassword123
              passwordConfirm: securepassword123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registration successful
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: User login
      description: Authenticates user and returns JWT tokens
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: securepassword123
              remember: false
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/TokenPair'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account is locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Uses refresh token to obtain a new access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token refreshed successfully
                  tokens:
                    $ref: '#/components/schemas/TokenPair'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: User logout
      description: Invalidates the refresh token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
        '400':
          description: Validation error or logout failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    get:
      summary: Get user profile
      description: Returns the authenticated user's profile information
      tags:
        - Profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update user profile
      description: Updates the authenticated user's profile information
      tags:
        - Profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: John Doe Updated
              bio: Software Developer
              avatar: https://example.com/avatar.jpg
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/change-password:
    put:
      summary: Change password
      description: Changes the authenticated user's password
      tags:
        - Profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              oldPassword: currentpassword123
              newPassword: newsecurepassword456
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password changed successfully
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized or invalid old password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - BearerAuth: [] 